# app/public/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "VeriBits API",
    "description": "File verification, security tools, and cryptographic validation platform API",
    "version": "1.0.0",
    "contact": {
      "name": "VeriBits Support",
      "email": "support@veribits.com",
      "url": "https://veribits.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://veribits.com/terms"
    }
  },
  "servers": [
    {
      "url": "https://veribits.com/api/v1",
      "description": "Production API"
    },
    {
      "url": "https://api.veribits.com/api/v1",
      "description": "Production API (alternate)"
    }
  ],
...
  "paths": {
...
    "/webhooks": {
      "get": {
        "tags": ["Webhooks"],
        "summary": "List webhooks",
        "operationId": "listWebhooks",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "List of webhooks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {"$ref": "#/components/schemas/Webhook"}
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Webhooks"],
        "summary": "Create webhook",
        "operationId": "createWebhook",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Webhook created",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Webhook"}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
...
      "Webhook": {
        "type": "object",
        "properties": {
          "id": {"type": "string"},
          "url": {"type": "string", "format": "uri"},
          "events": {"type": "array", "items": {"type": "string"}},
          "active": {"type": "boolean"}
        }
      },
      "WebhookCreate": {
        "type": "object",
        "required": ["url", "events"],
        "properties": {
          "url": {"type": "string", "format": "uri"},
          "events": {"type": "array", "items": {"type": "string"}},
          "secret": {"type": "string"}
        }
      }
    },
...
  }
}